myproject/requirements.txt
Django==4.0
djangorestframework==3.13.1
pytest==7.1.2
pytest-django==4.5.2
